fastuidraw/painter/packing: use delayed deletion of resources
Use the delayed deletion interface of ImageAtlas and ColorStopAtlas
in PainterPacker so that if an Image or ColorStopSequenceOnAtlas
is used by a PainterBrush and is deleted, its actual deletion
is delayed until after they are no longer needed. Note that
PainterBrush::derived_value_type still tracks what Image
and ColorStopSequenceOnAtlas are used. This is because the
Image and ColorStopSequenceOnAtlas need to NOT be deleted as
long as a PainterPackedValue<PainterBrush> made from such a
PainterBrush are still alive.